Package into duke package and subpackages
