+++ src/main/java/duke/Duke.java
@@ -1,12 +1,7 @@
-import java.io.IOException;
+package duke;
 
-/**
- * A Personal Assistant Chatbot that helps a person to keep track of various things.
- * CS2103T iP
- * AY22/23 Semester 1
- *
- * @author Perry Wong
- */
+import java.io.IOException;
+import duke.command.Command;
 
 public class Duke {
