+++ src/main/java/duke/Echo.java
@@ -1,19 +1,9 @@
-/**
- * Represents an echo function which echoes whatever the user inputs.
- * CS2103T iP
- * AY22/23 Semester 1
- *
- * @author Perry Wong
- */
-
-// Level 1
+package duke;
 
 import java.util.Scanner;
 
 public class Echo {
-    /**
-     * The method to echo the user's input.
-     */
+
     public static void echo() {
         Scanner sc = new Scanner(System.in);
         String str = sc.nextLine();
