+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.task.Task;
 import java.util.ArrayList;
 
 public class TaskList {
