+++ src/main/java/duke/command/Command.java
@@ -1,4 +1,10 @@
+package duke.command;
+
 import java.io.IOException;
+import duke.Ui;
+import duke.Storage;
+import duke.TaskList;
+import duke.DukeException;
 
 public abstract class Command {
