+++ src/main/java/duke/command/MarkCommand.java
@@ -1,4 +1,11 @@
+package duke.command;
+
 import java.util.ArrayList;
+import duke.Ui;
+import duke.Storage;
+import duke.TaskList;
+import duke.DukeException;
+import duke.task.Task;
 
 public class MarkCommand extends Command {
     private int index;
@@ -14,7 +21,7 @@ public class MarkCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         ArrayList<Task> list = taskList.getTaskArrayList();
-        if ((index > 0) && (list.get(index - 1) != null)) {
+        if ((index > 0) && index < list.size() && (list.get(index - 1) != null)) {
             Task t = list.get(index - 1);
             taskList.mark(this.index);
             ui.showMark(t);
