+++ src/main/java/duke/command/ShowListCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Ui;
+import duke.Storage;
+import duke.TaskList;
+
 public class ShowListCommand extends Command {
     @Override
     public boolean isExit() {
