+++ src/main/java/duke/task/Task.java
@@ -1,0 +1,29 @@
+package duke.task;
+
+public class Task {
+
+    protected String description;
+
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void mark() {
+        this.isDone = true;
+    }
+
+    public void unmark() {
+        this.isDone = false;
+    }
+
+    public String toString() {
+        return "[" + getStatusIcon() + "] " + description;
+    }
+}
