+++ src/main/java/duke/task/Todo.java
@@ -1,0 +1,13 @@
+package duke.task;
+
+public class Todo extends Task {
+
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
